backend/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
numpy
sentence-transformers
torch
pydantic
pandas
kagglehub
firebase-admin
